+++ packages/executor/src/actions/index.ts
@@ -4,6 +4,5 @@ export * from "./createStateOverride";
 export * from "./finalizeBTCTransaction";
 export * from "./getBTCFeeRate";
 export * from "./getPublicKey";
-export * from "./getPublicKeyForAccount";
 export * from "./signIntention";
 export * from "./signTransaction";