+++ packages/satoshi-kit/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @midl-xyz/satoshi-kit
 
+## 2.0.0-next.49
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-react@2.0.0-next.49
+
+## 2.0.0-next.48
+
+### Patch Changes
+
+- feat(satoshi-kit): expose account and network to verify and sign message
+  feat(executor): getEVMAddress is now possible to use w/o midl config
+
 ## 2.0.0-next.47
 
 ### Patch Changes